classify exception
not 100% sure on this one
